calculate yuv compress time and yuv decompress time

+++ include/tj_test.h
@@ -16,5 +16,5 @@ typedef struct tjp_info {
 int get_timer_now ();
 int tj_test();
 uchar *read_file2buffer(char *filepath, tjp_info_t *tinfo);
-void write_buffer2file(char *filename, uchar *buffer, int size);
+void write_buffer2file(char *filename, uchar *buffer, unsigned long size);
 #endif //TEST_JPG_TURBO_TJ_TEST_H
